+++ api/src/main/java/it/unicam/cs/pa/ma114763/logo/Point.java
@@ -4,21 +4,21 @@ package it.unicam.cs.pa.ma114763.logo;
  * @author Lorenzo Lapucci
  */
 public class Point implements Position2D {
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     @Override
-    public int getX() {
+    public double getX() {
         return x;
     }
 
     @Override
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
